refactor(config): keep output projections in the model layer
